Finish stomping memory-related bugs in Get trackless quality michels.
Remove MichelEvent.h. Move contents into Michel.h to avoid include complications.

Avoid some forward declares.

The bug fixes: add a DTOR to MichelEvent. Don't push back into
a vector of pointers inside of a loop using &.
